Navbar links work well + all tests pass
now navbar fetch link in generic way - one call to a specific function
in views.

all tests pass

added the example user to conftest
