+++ AcadeMeData/tests/test_client.py
@@ -1,16 +1,13 @@
 import pytest
 
-from AcadeMe import views
-
 
 @pytest.mark.django_db
 def test_homepage(client):
-    request = client.get('')
-    response = views.homePage(request)
+    response = client.get('')
     assert response.status_code == 200
 
 
+@pytest.mark.django_db
 def test_contactus(client):
-    request = client.get('contact_us/')
-    response = views.contact(request)
+    response = client.get('/contact_us/')
     assert response.status_code == 200
