+++ templates/app_layout.html
@@ -16,41 +16,22 @@
     <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
 
 </head>
+
 <body>
 <nav class="navbar navbar-expand-sm">
     <div
             class="collapse navbar-collapse d-flex justify-content-between"
-            id="collapsibleNavbar"
-    >
+            id="collapsibleNavbar">
 
         <ul class="navbar-nav">
             <li>
                 <a class="nav-link mr-4 app-title" href="/">AcadeMe</a>
-
             </li>
-            {% for link_title, link_address in navbar_links %}
+            {% for link_title, link_address in navbar_links.items %}
             <li class="nav-item">
-                <a class="nav-link mr-4" href={{ link_address }}>{{ link_title }}</a>
+                <a class="nav-link mr-4" href={{ link_address }}> {{ link_title }} </a>
             </li>
             {% endfor %}
-
-
-            <!--            {% if not user.is_authenticated %}-->
-            <!--            <li>-->
-            <!--                <a class="nav-link mr-4" href="/login">Log In</a>-->
-            <!--            </li>-->
-            <!--            <li>-->
-            <!--                <a class="nav-link mr-4" href="/register">Register</a>-->
-            <!--            </li>-->
-            <!--            {% else %}-->
-            <!--            <li>-->
-            <!--                <a class="nav-link mr-4" >Welcome {{ user.username }} </a>-->
-            <!--            </li>-->
-            <!--            <li>-->
-            <!--                <a class="nav-link mr-4" href="/logout">Log Out</a>-->
-            <!--            </li>-->
-            <!--            {% endif %}-->
-
         </ul>
         <ul class="navbar-nav">
             <li>
@@ -64,4 +45,5 @@
     {% block content %} {% endblock %}
 </div>
 </body>
+
 </html>
